Modularize shared chat dock
